initial integration of HTTP with MQTT and motion detection (#27)
* initial integration of HTTP with MQTT and motion detection

* PR redundancy testing and rework #1

* wifi config error fix and delay adjust

* PR nits

This PR appends core ESP-IDF wifi functionality in configuring and creating an access point (AP) to facilitate camera streaming to the browser. This is "appended" post initialization of the WiFi libraries and setup of the ESP EYE as a station STA. This is a workaround for integrating HTTP streaming and MQTT publishing considering the AFR wifi functions do not facilitate operation as both an AP and as a station (STA), which is the mode required to publish MQTT messages. Further work is required to either fork and alter the MQTT AFR implementation to include AP & STA operation OR to replace MQTT's AFR use with core ESP-IDF wifi functions which allow for simultaneous modes of operation.